+++ eCV/reconstruction/ProjectWidget.h
@@ -11,10 +11,11 @@
 #include <QtWidgets>
 
 #include "util/misc.h"
+#include "util/option_manager.h"
 
 namespace cloudViewer {
 
-class OptionManager;
+using OptionManager = colmap::OptionManager;
 
 class ProjectWidget : public QWidget {
 public: